requirements.txt
flake8
pycodestyle
mypy
ruff
pylint
loguru~=0.7.2
pytest~=8.0.0
robotframework~=7.0.1
